Refactor jwt auth headers and cookie with new changes
