+++ setup.cfg
@@ -53,9 +53,6 @@ per-file-ignores =
   dmr/throttling/*.py: WPS226
   # It is fine to have many exceptions:
   dmr/exceptions.py: WPS202
-  # Base jwt auth, its header transport, and the request helpers
-  # all have to live here, so the old import paths keep working:
-  dmr/security/jwt/auth.py: WPS202
   # Auth base classes live here together with the response spec
   # and `WWW-Authenticate` helpers that all of them share:
   dmr/security/base.py: WPS202
